# server/api-docs.yaml
openapi: "3.0.2"
info:
  title: Nested REST APIs
  version: "1.0"
servers:
  - url: https://api.server.test/api/v1
paths:
...
  /children:
    get:
      summary: Get all children
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Child'
    post:
      summary: Create a new child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/Child'
      responses:
        '201':
          description: Child created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Child'
    put:
      summary: Update an existing child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/Child'
      responses:
        '200':
          description: Child updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Child'
...
  /children/{child_id}/actions/place_order:
    post:
      summary: Place orders for all active goals
      requestBody:
        description: Orders placed
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/PlaceOrderPost'
      responses:
        '201':
          description: Orders placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlaceOrder'
  /children/{child_id}/actions/verifiy_order:
    post:
      summary: Verify placed orders
      requestBody:
        description: Verification details
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                verification_code:
                  type: string
      responses:
        '200':
          description: Orders verified
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/PlaceOrder'

        '400':
          description: Orders verification failed
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/PlaceOrder'
...
components:
  schemas:
    Child:
      type: object
      required:
        - first_name
        - last_name
        - dob
        - gender
        - invest_under_child
      properties:
        id:
          type: string
          format: number
        first_name:
          type: string
        last_name:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum:
            - male
            - female
            - other
        invest_under_child:
          type: boolean
        user:
          type: object
          $ref: '#/components/schemas/MinifiedUser'
...
    MinifiedUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    PlaceOrder:
      type: object
      properties:
        id:
          type: string
          format: number
        verification_status:
          type: string
          enum:
            - pending
            - verified
            - failed
        status:
          type: string
          enum:
            - pending
            - completed
            - failed
            - cancelled
        payment_url:
          type: string
          format: url
        mandate:
          type: object
          properties:
            type:
              type: string
              enum:
                - net_banking
                - upi
            upi_id:
              type: string
            confirmation_url:
              type: string
    PlaceOrderPost:
      type: object
      properties:
        orders:
          type: array
          items:
            type: object
            properties:
              goal:
                type: object
                $ref: '#/components/schemas/MinifiedGoal'
              BuyOrder:
                type: object
                $ref: '#/components/schemas/BuyOrder'
              SipOrder:
                type: object
                $ref: '#/components/schemas/SipOrder'
        payment_method:
          type: string
          enum:
            - net_banking
            - upi
    BuyOrder:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: float
    SipOrder:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: float
        setup_option:
          type: object
          required:
            - amount
            - frequency
          properties:
            amount:
              type: number
              format: float
            frequency:
              type: string
              enum:
                - monthly
                - quarterly
                - half-yearly
                - yearly
    MinifiedGoal:
      type: object
      properties:
        id:
          type: string
          format: number
...